backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
django-filter>=23.5
psycopg2-binary>=2.9
python-dotenv>=1.0
anthropic>=0.18
gunicorn>=21.2
whitenoise>=6.6
stripe>=8.0
openpyxl>=3.1
